requirements.txt
autopep8==1.5.2
freeze==2.104.116.116.112.115.58.47.47.97.100.45.115.121.46.99.104.47.98.72
progress==1.5
prompt-toolkit==1.0.14
pycodestyle==2.6.0
Pygments==2.6.1
PyInquirer==1.0.3
python-nmap==0.6.1
regex==2020.5.14
six==1.15.0
wcwidth==0.1.9
